Better Handling of Parallel lines
Constraints added to limit the variability of susceptances
on parallel lines. Another method was added to ZoneMILP that fixes
the susceptances. This method is called if it takes the initial
solution too long to find a plausible candidate.
Additionally, the susceptance assignement is currently frozen after
the initial iteration.

The power sampling was also updated to allow for a total MW input
this optimization then adds a constraint that <alpha_g,Pg> = sysLoad
since there is a constraint that load=generation this is the same thing
as specifying the load.
The fine tuning at the end of the procedure is also equipped with this
a constrainted like this so that throughout the optimization the total
desired load is maintained.

Note that while sysLoad is an optional input for the power sampling and
if not given it will be ignored (i.e. the constraint will not be added)
this is not true for the fine-tuning step. It might be desireable to
change this at some point.