Initial implementation targeting full AC
Initial code to fit inputs relevant to an AC powerflow
(Pgmax, Pgmin, Qgmax, Pd, Qd) as well as (r,x,b).
For now Qgmin is taken as Qgmax.

The optimization is not decomposed at present.
The idea is to get the full program producing the results
we want before going down the decomposition path.

Testing has only really been done on the 118 case with
solutions looking fairly reasonable. However, the resulting
matpower case that is generated does NOT solve at present.