Loss and flexible fmax work
Fixed loss terms so they are (Pg-Pd)/Pg instead of (Pg-Pd)/Pd.
Starting work on allowing for flexible fmax, that is, include a line
limit as one of the branch parameters. This should be included correctly
in the data structure resz coming out of fit_inputs.py but has not been
otherwise integrated into the rest of the code base.