Changes to Big M and Line Ratings
Continuing integration of line rating into formulation. The code should
now take both an fmax argument as well as consider the line ratings in
the data. During the sampling process in helpers.py an ACTUAL fmax is
determined as the maximum of the input fmax and the maximum value in the
data. All zeros in the data are replaced with this value.

Line flows are therefore now initialized unconstrained and equations are
used to bound them. There is a bit of code that tires to determine
wether there are different line ratings or just 1. If only 1 is found
then the variable is bound.

Similar changes are made to multvar_tieline.py

The bigM calculation has also been modified. it now:
a) consideres the effect of phi, which was previously neglected.
b) save 4 different values for pf, pt, qf, pt, which are then used
rather than taking the maximum of the four.
relevant logging scripts have also been changed.