Bus Shunt Variables and reworking of line ratings
Shunts are added to the formulation as slack variables.
The main idea is to minimize the amount of shunts
but at the same time enable their use to keep voltage levels
in check.
CAUTION: It appears there are some issues here and for the moment
mosts tests are done neglecting these shunts.

The line parameters have also been reworked a bit so that now
an option can be passed during the fitting stage as to whether
a single, constant line limit should be used.