Branch Flow Equation fix
There was an error in the branch flow constraints (sign errors).
Additionally, the Q constraints incorrectly added phi instead of
subtracting it.
These should now be fixed. All angle differences have now been set
to from-to for consitency as well.
The solution check routine was also changed accordingly.

Finally, the bigM calculation function was updated since a few
considerations were not correctly considred initially.
This is better documented and derived in the PlacementExpansion
document.