Tie-Line and bigM fixes
Added umin as argument to the various bigM function calls, since this value
is now needed when calculating bigM.
Went through the tie-line optimization and brought it up to date:
- shunts added
- correction to powerflow formultation (from-to for angles and errors in
  Q)
- correction to how rates are handled.

NOTE: the way rates were handled does not seem logical. It appears that
the ztie dictionary which has vectors of length E was indexed with
global line ids, which should have been out of range. This suggests that
either gurobi silently ignored the error or that somehow there is a
problem with how the if statement is formulated (possibly with tset).
This should be looked into at some point.

Also added an if statement in the objective of the single system
formulation in formulation_multvar.py. This is needed for when no shunts
are used, and seems to have been left out previously.