Updating mpc output
The make mpc function has been updated to allow it to handle the data
created by the EA approach, i.e. multiple individuals.
This is effectively done by a single recursive call if the field 'vars'
is in the data.

The function can now also receive data directly rather than just a
string where the data is stored. The idea is to actually incorportate
this directly into the run file.