Modifiable rho and error oriented polyhedral relax
a method to modify rho is added to the multvar_solve.py functions
This way rho can be slowly reduced, which might help prevent oscillating
behavior.

The number of hyperplances used for the polyhedral relaxation of squared
terms is now calculated using the desired maximum error (function in
helpers). The one danger is that in some cases this can yield a very
large constraint set.

Logging updated to show the error between the squared terms and their
polyhedral relaxed variables. This is predominantely due to the fact
that phi is sometimes larger than it needs to be despite the
minimization, because it is is somehow confilicting with the desire to
minimize Pg. The callback function now keeps track of these violations,
mainly for informative reasons. The attempt to force better agreement in
this area was not very successful and is probably not of highest
concern.