Parallel zone implementation
Individual zones can now be run in parallel, which will hopefully
further speed things up. Because these are child processes of the
individuals, these needed to be non daemonic processes. Therefore, the
implementation of the individual parallellization had to chage as well
since Pool automatically makes daemonic spawns.

Finally, there is a limit of 5 individuals running at a time to
hopefully avoid out of memory issues.