Slack variable for Qg
In a continuing effor to force Qg to be non negative but maintain
feasibility, a slack variable so the Qg total constraint is now:
                    sum(Qg) + Qgslack >= 0
Where Qgslack >=0 and is added to the objective with a fairly large
penalty.

This has yet to finish, but it did go a few generations, so the outlook
is promising.