Different Topology, and random assignment
multvar_init.py now has some functions to generate different graphs
(currently only Erdos Reney, ER, graphs are implemented).
ea_run.py takes an optional 'topology' argument to determine whether the
case that provides the load and impedance data should also be used for
topology or whether a different topology should be generated.

Separately an option to randomly place load parameters as well has been
added. This option:
- automatically sets Nmax so there will be only 1 zone
- generates a random permutation and sets self._S{} so that the a simple
  linear program will be inialized

In some cases the problem becomes infeasible (this was also occasionally
true previously). Some more handling is added so in the event of
infeasibility or unboudedness the objective is set to np.inf, which
should mean the model will be tossed out during the selection process
(provided enough models solved...)

Finally, an absolute value variable was added to the formulation to keep
track of the magnitude of the reactive power flows.
These are then added to the minimization objective.
This is done since from observation the reactive flows were consistently
a bit too high.