Modification to line flow handling
1) line derating:
Instead of limiting both P and Q flows to be below the line rating,
flows are now limited to be below rating/sqrt(2). This conservative
derating should hopefully make the AC OPF solve without having to adjust
ratings any further than already done with slack variables.

2) SIL average
The SIL for each branch that is not a transformer is calculated in a
function in helpers calc_sil().
This function returns a dictionary indexed like the zparameter
dictionary except that entries which correspond to transformers are
removed.
During the edge constraints loop, the Pf of branches in the sil
dictionary are added to a LinExpr that then forms a constraint:
	c/L*sum(Pfabs/SIL) + ssil <= 1
c is a factor to scale from the fact that we use Pf to actually wanting
Sf, L is the number of branches uses i.e. we average (Pf/SIL).
Both Pfabs (out of necessity) and ssil are added to the objective.