Adjustment to zone_splitting algorithm
Previously, when a component was created that was too small it was
immediately attached to a neighboring component.
This lead to a situation where the zone which is too small would be
attached to the largest zone. Then on the subsequent iteration it
is simply split away from that zone again leading to an endless
loop.

Two changes are made:
1) Nminflag is a boolean which is set any time a component is found
which is too small. On the iteration after Nminflag is set to True
the algorithm skips checking for small components.
This way, the largest component will be broken in 2 and possibly
the small component that will be attached on the subsequent round
will either a) not be broken off, or b) not exceed Nmax

2) instead of picking the first neighbor component all neighbor
components are collected in a dictionary and then the neighbor
with the FEWEST nodes is chosen. This is again done in the hope
of not exceeding Nmax when possible.