Some error handling and debugging
In ea.py another try, catch statement is added in parzone to avoid
infeasible solutions being sent to the parent process and then hanging
I don't know why this is happening, frankly, the previous try
statement should catch it, but that's what appears to be needed.

in multvar_solution_check there is a slight change to how the script
checks whether there are any shunt reactances in the data.
Some more stuff is added to the __init__="__main__" section,
so that the script can be called on saved data at a later time.