Generation Qlimits
Some generators have very large Qlimits (essentiall the equivalent of
inifinit). I suspect that when these are multiplying the Pi permutation
matrix it is causing all sorts of numerical issues, which result in
infeasibility or unboundedness.

This is based on an observation that the zones that appeared to
constantly be returning problems were the ones that had such infinit
limits while the others were not. Far from conclusive, but semi
convincing.

In formulation_ea a variable is created called Qgmax, which is a copy of
the input params['S']['Qgmax']. Then every entry that is larger than
twice the sum of all the reactive load in the zone
(2*params['S']['Qd'].sum()) is replaced by 2*params['S']['Qd'].sum().

Qgmax is then used in creating the constraints, however it is NOT used
in exporting the variables, that way the final solution should still
have the "correct" limits