Adding generator cost to fit and sample
When loading data the function in helpers now returns 4 panda's
dataframes the last of which includes generator cost.
When fitting the data, the cost associated with generators at a given
bus is the weighted average of those generators: sum(c_i*P_i)/sum(P_i).

The cost is also part of the samples generated.

Unfortunately, including the costs in the problem formulation is
problematic since it creates bilinear terms (the costs must be permuted
with matrix Pi). Therefore this change should really make a difference,
there will be more data passed arround, but it will actually not be
used.
 Changes to be committed:
	modified:   ea_run.py
	modified:   fit_inputs.py
	modified:   helpers.py